+++ Package.swift
@@ -1,0 +1,29 @@
+// swift-tools-version:5.3
+
+import PackageDescription
+
+let package = Package(
+    name: "SDOSCustomLoader",
+    platforms: [
+        .iOS(.v9)
+    ],
+    products: [
+        .library(
+            name: "SDOSCustomLoader",
+            targets: ["SDOSCustomLoader"])
+    ],
+    dependencies: [
+    ],
+    targets: [
+        .target(
+            name: "SDOSCustomLoader",
+            path: "src",
+            publicHeadersPath: "",
+            cSettings: [
+                .headerSearchPath("Classes"),
+                .headerSearchPath("Classes/Helper"),
+                .headerSearchPath("Classes/Progress"),
+                  ]
+        ),
+    ]
+)